[#4] Refactor: 낙관적 락 Facade 메소드의 분리 구현
