+++ src/main/java/com/shj/onlinememospringproject/service/MemoFacade.java
@@ -1,0 +1,7 @@
+package com.shj.onlinememospringproject.service;
+
+import com.shj.onlinememospringproject.dto.MemoDto;
+
+public interface MemoFacade {
+    void updateMemoFacade(Long memoId, MemoDto.UpdateRequest updateRequestDto);
+}
